requirements.txt
alembic==1.3.1
cachetools==3.1.1
certifi==2019.11.28
chardet==3.0.4
Click==7.0
configparser==4.0.2
Flask==1.1.1
flask-marshmallow==0.10.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
google-api-python-client==1.7.11
google-auth==1.7.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
httplib2==0.14.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.1.0
MarkupSafe==1.1.1
marshmallow==3.2.2
marshmallow-sqlalchemy==0.20.0
mysql-connector==2.2.9
oauthlib==3.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.13.0
SQLAlchemy==1.3.11
uritemplate==3.0.0
urllib3==1.25.7
Werkzeug==0.16.0
